+++ src/gradcam.jl
@@ -15,16 +15,25 @@ GradCAM is compatible with a wide variety of CNN model-families.
 # References
 - $REF_SELVARAJU_GRADCAM
 """
-struct GradCAM{F,A} <: AbstractXAIMethod
+struct GradCAM{F,A,B<:AbstractADType} <: AbstractXAIMethod
     feature_layers::F
     adaptation_layers::A
+    backend::B
+
+    function GradCAM(
+        feature_layers::F, adaptation_layers::A, backend::B=DEFAULT_AD_BACKEND
+    ) where {F,A,B<:AbstractADType}
+        new{F,A,B}(feature_layers, adaptation_layers, backend)
+    end
 end
 function call_analyzer(input, analyzer::GradCAM, ns::AbstractOutputSelector; kwargs...)
     A = analyzer.feature_layers(input)  # feature map
     feature_map_size = size(A, 1) * size(A, 2)
 
     # Determine neuron importance αₖᶜ = 1/Z * ∑ᵢ ∑ⱼ ∂yᶜ / ∂Aᵢⱼᵏ
-    grad, output, output_indices = gradient_wrt_input(analyzer.adaptation_layers, A, ns)
+    grad, output, output_indices = gradient_wrt_input(
+        analyzer.adaptation_layers, A, ns, analyzer.backend
+    )
     αᶜ = sum(grad; dims=(1, 2)) / feature_map_size
     Lᶜ = max.(sum(αᶜ .* A; dims=3), 0)
     return Explanation(Lᶜ, input, output, output_indices, :GradCAM, :cam, nothing)