+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # ExplainableAI.jl
+## Version `v0.9.0`
+- ![Feature][badge-feature] Support selection of AD backend via DifferentiationInterface.jl ([#167])
+  - `Gradient`, `InputTimesGradient` and `GradCAM` analyzers now have an additional `backend` field and type parameter
+- ![Maintenance][badge-maintenance] Update XAIBase interface to v4 ([#166])
+
 ## Version `v0.8.0`
 This release removes the automatic reexport of heatmapping functionality.
 Users are now required to manually load 
@@ -210,6 +215,8 @@ Performance improvements:
 [VisionHeatmaps]: https://julia-xai.github.io/XAIDocs/VisionHeatmaps/stable/
 [TextHeatmaps]: https://julia-xai.github.io/XAIDocs/TextHeatmaps/stable/
 
+[#167]: https://github.com/Julia-XAI/ExplainableAI.jl/pull/167
+[#166]: https://github.com/Julia-XAI/ExplainableAI.jl/pull/166
 [#162]: https://github.com/Julia-XAI/ExplainableAI.jl/pull/162
 [#159]: https://github.com/Julia-XAI/ExplainableAI.jl/pull/159
 [#157]: https://github.com/Julia-XAI/ExplainableAI.jl/pull/157